requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
streamlit
requests
joblib
paho-mqtt
